requirements/requirements.txt
Django==4.0.2
django-allauth==0.48.0
django-compressor==3.1
django-environ==0.8.1
django-extensions==3.1.5
django-js-asset==1.2.2
django-modeltranslation==0.17.5
django-countries[pyuca]==7.3.2
django-admin-csvexport==1.11
ipython==8.0.1
Pillow==9.0.1
psycopg2-binary==2.9.3
pydash==5.1.0
requests==2.27.1
uWSGI==2.0.18
